folder restructure, filter UI code
