+++ paystatsapp-ui/package.json
@@ -8,8 +8,10 @@
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.21.0",
     "bootstrap": "^4.5.2",
+    "date-fns": "^2.16.1",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
+    "react-date-range": "^1.1.3",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.3"
   },
