+++ paystatsapp-ui/src/MainContainer.css
@@ -29,8 +29,10 @@
     font-weight: normal;
 }
 
-.rounded-button{
-    border-radius: 15px !important;
+.show-pointer{
+    cursor: pointer;
 }
+
+
