+++ paystatsapp-ui/src/payments/ViewPayments.css
@@ -1,0 +1,3 @@
+.rounded-button{
+    border-radius: 15px !important;
+}
